プリントレイアウトのメインバー隠し処理をmodules配下のtags.jsに移した #297
